backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
nltk==3.9.1
numpy==2.2.3
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
six==1.17.0
tqdm==4.67.1
tzdata==2025.1
Werkzeug==3.1.3
